+++ lib/MacroTest/src/MacroTest.cpp
@@ -1,0 +1,11 @@
+#include "MacroTest.h"
+
+Library::Library(){}
+
+#if PS_DEBUG 
+
+    void Library::print(){
+        Serial.println("Printing");
+    }
+
+#endif // DEBUG