+++ src/main.cpp
@@ -1,68 +1,8 @@
-#include <Arduino.h>
-#include <HardwareSerial.h>
+
+#define PS_DEBUG false
+
 #include <PacketSerial.h>
 
-
-
-
-    // #define __PACKET_SERIAL_CONFIG__
-
-        
-    //     /* Comment out this next define if your MCU is not an ESP32.
-    //     *  The ESP32 platform includes FreeRTOS and the FreeRTOS libraries do
-    //     *  not have to be loaded seperately.*/
-    //     #define __PLATFORM_IS_ESP32__
-
-    //     /* Comment out this next define to silence debug printing to the
-    //     * terminal.
-    //     */
-    //     /* Uncomment this define to use software serial. */
-    //     // #define PS_USE_SOFTWARE_SERIAL
-
-    //     #if PS_USE_SOFTWARE_SERIAL
-    //         #include <SoftwareSerial.h>             // featherfly/SoftwareSerial@^1.0
-    //     #else
-    //         #include <HardwareSerial.h>
-    //         // #define PS_HWS_RX_GPIO 16               // The RX GPIO.
-    //         // #define PS_HWS_TX_GPIO 17               // The TX GPIO.
-    //         // #define PS_HWS_INVERT_LOGIC false       // Invert the pin logic to active high.
-    //     #endif // PS_USE_SOFTWARE_SERIAL
-        
-    //     /// @brief The default serial port speed.
-    //     #define PS_BAUD 115200
-
-    //     /// @brief The processor core that runs the serial port processes.
-    //     #define PS_CORE 0
-
-    //     /// @brief The name of the serial port RX task.
-    //     #define PS_RX_TASK_NAME "PS_RX_TASK"
-
-    //     /// @brief The name of the serial port TX task.
-    //     #define PS_TX_TASK_NAME "PS_TX_TASK"
-
-    //     /// @brief The priority of the serial monitor tasks.
-    //     #define PS_TASK_PRIORITY 1
-
-    //     /// @brief The stack size for the serial port RX task.
-    //     #define PS_RX_STACK_SIZE 10800
-
-    //     /// @brief The stack size for the serial port TX task
-    //     #define PS_TX_STACK_SIZE 10800
-
-    //     /// @brief The length of the RX queue.
-    //     /// 
-    //     /// Increasing the queue length may require an increase in stack size. 
-    //     #define PS_RX_QUEUE_LENGTH 5
-
-    //     /// @brief The length of the TX queue.
-    //     /// 
-    //     /// Increasing the queue length may require an increase in stack size. 
-    //     #define PS_TX_QUEUE_LENGTH 5
-
-    // /// @brief The length of the ERROR queue.
-    //     /// 
-    //     /// Increasing the queue length may require an increase in stack size. 
-    //     #define PS_ERR_QUEUE_LENGTH 256
 
 /// @brief The valid headers that can be sent by the device.
 static const std::vector<uint16_t> headers{0x5aa5};
@@ -92,7 +32,18 @@ PacketSerial display = PacketSerial(headers, & displayPort);
 uint16_t heading = -1;
 
 void getHeading(){
+  
+  // #if PS_DEBUG 
+
+  //   lib.print();  
+
+  // #endif      
   ps_frame_t frame = display.read();
+  #if PS_DEBUG
+    Serial.print("Frame: { ");
+    frame.print();
+    Serial.println(" }");
+  #endif //PS_DEBUG
   uint16_t address = ((frame.data[1]) << 8) | frame.data[2];
   if (address == 0x5000){
      heading = uint16_t(360 - (((frame.data[4]) << 8) | frame.data[5]) / 2);
@@ -156,7 +107,9 @@ void setup() {
 void loop() {
   
   // long delay alows testing of exception handler
-  delay(50);                          
+  delay(50);  
+
+                    
   
   uint8_t error = 0xff;                 // placeholder for errors